chore: move luhncheck function
